refactor: JdbcTemplate를 jdbc 패키지로 이동시킨다.
