--- app/src/main/java/com/techcourse/dao/JdbcTemplate.java
@@ -1,72 +1,0 @@
-package com.techcourse.dao;
-
-import java.sql.Connection;
-import java.sql.PreparedStatement;
-import java.sql.ResultSet;
-import java.sql.SQLException;
-import java.util.ArrayList;
-import java.util.List;
-import javax.sql.DataSource;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-
-public abstract class JdbcTemplate {
-
-    protected static final Logger log = LoggerFactory.getLogger(JdbcTemplate.class);
-
-    private final DataSource dataSource;
-
-    public JdbcTemplate(DataSource dataSource) {
-        this.dataSource = dataSource;
-    }
-
-    public void update(String sql, PreparedStatementSetter preparedStatementSetter) throws SQLException {
-        Connection conn = dataSource.getConnection();
-        PreparedStatement pstmt = conn.prepareStatement(sql);
-
-        try (conn; pstmt) {
-
-            preparedStatementSetter.setValues(pstmt);
-            pstmt.executeUpdate();
-        }
-    }
-
-    private ResultSet executeQuery(PreparedStatement pstmt) throws SQLException {
-        return pstmt.executeQuery();
-    }
-
-    public Object queryForObject(String sql, PreparedStatementSetter preparedStatementSetter, RowMapper rowMapper) throws SQLException {
-        Connection conn = dataSource.getConnection();
-        PreparedStatement pstmt = conn.prepareStatement(sql);
-        preparedStatementSetter.setValues(pstmt);
-        ResultSet rs = executeQuery(pstmt);
-
-        try (conn; pstmt; rs) {
-            log.debug("query : {}", sql);
-
-            if (rs.next()) {
-                return rowMapper.mapRow(rs);
-            }
-            return null;
-        }
-    }
-
-    public Object query(String sql, PreparedStatementSetter preparedStatementSetter, RowMapper rowMapper) throws SQLException {
-        List<Object> results = new ArrayList<>();
-
-        Connection conn = dataSource.getConnection();
-        PreparedStatement pstmt = conn.prepareStatement(sql);
-        preparedStatementSetter.setValues(pstmt);
-        ResultSet rs = executeQuery(pstmt);
-
-        try (conn; pstmt; rs) {
-            log.debug("query : {}", sql);
-
-            if (rs.next()) {
-                results.add(rowMapper.mapRow(rs));
-            }
-            return results;
-        }
-    }
-
-}
